test(wallet): refactor integration tests
