+++ packages/wallet/test/integration/transactionTime.test.ts
@@ -1,37 +1,13 @@
-import { Cardano, createSlotTimeCalc, testnetTimeSettings } from '@cardano-sdk/core';
-import { KeyManagement, SingleAddressWallet, SingleAddressWalletProps } from '../../src';
-import { createStubStakePoolSearchProvider, createStubTimeSettingsProvider } from '@cardano-sdk/util-dev';
+import { SingleAddressWallet } from '../../src';
+import { createSlotTimeCalc } from '@cardano-sdk/core';
+import { createWallet } from './util';
 import { firstValueFrom } from 'rxjs';
-import { mockAssetProvider, mockTxSubmitProvider, mockWalletProvider } from '../mocks';
-
-const walletProps: SingleAddressWalletProps = { name: 'some-wallet' };
-const networkId = Cardano.NetworkId.mainnet;
-const mnemonicWords = KeyManagement.util.generateMnemonicWords();
-const getPassword = async () => Buffer.from('your_password');
 
 describe('integration/transactionTime', () => {
-  let keyAgent: KeyManagement.KeyAgent;
   let wallet: SingleAddressWallet;
 
   beforeAll(async () => {
-    keyAgent = await KeyManagement.InMemoryKeyAgent.fromBip39MnemonicWords({
-      getPassword,
-      mnemonicWords,
-      networkId
-    });
-    const txSubmitProvider = mockTxSubmitProvider();
-    const walletProvider = mockWalletProvider();
-    const stakePoolSearchProvider = createStubStakePoolSearchProvider();
-    const timeSettingsProvider = createStubTimeSettingsProvider(testnetTimeSettings);
-    const assetProvider = mockAssetProvider();
-    wallet = new SingleAddressWallet(walletProps, {
-      assetProvider,
-      keyAgent,
-      stakePoolSearchProvider,
-      timeSettingsProvider,
-      txSubmitProvider,
-      walletProvider
-    });
+    wallet = await createWallet();
   });
 
   it('provides utils necessary for computing transaction time', async () => {
